debug: `create_dwarf_cache` -> `.elf`
