+++ debug/elf.py
@@ -2,7 +2,12 @@
 
 __all__ = [
     "InMemoryELFFile"
+  , "create_dwarf_cache"
 ]
+
+from .dic import (
+    DWARFInfoCache,
+)
 
 from elftools.elf.elffile import (
     ELFFile
@@ -38,3 +43,22 @@ class InMemoryELFFile(ELFFile):
                     break
 
         super(InMemoryELFFile, self).__init__(stream)
+
+
+def create_dwarf_cache(exec_file):
+    elf = InMemoryELFFile(exec_file)
+    if not elf.has_dwarf_info():
+        raise ValueError(
+"%s does not have DWARF info. Provide a debug build\n" % (exec_file)
+        )
+
+    di = elf.get_dwarf_info()
+
+    if not di.debug_pubnames_sec:
+        print("%s does not contain .debug_pubtypes section. Provide"
+            " -gpubnames flag to the compiler" % exec_file
+        )
+
+    return DWARFInfoCache(di,
+        symtab = elf.get_section_by_name(".symtab")
+    )
